issue #4 - ShotAssistant - display minValue param

+++ src/main/java/com/gitlab/daring/image/assistant/PositionLimits.java
@@ -10,7 +10,7 @@ class PositionLimits extends BaseComponent {
 	final int offset = config.getInt("offset");
 	final double scale = config.getDouble("scale");
 	final double angle = config.getDouble("angle");
-	final double minValue = config.getDouble("minValue");
+	final int minValue = config.getInt("minValue");
 
 	PositionLimits(Config config) {
 		super(config);
